requirements.txt
django
djangorestframework
environs
djangorestframework-simplejwt
drf-spectacular
openapi-generator-cli